feat(nomina): implement personalized withdrawal service and UI
- Added RetiroPersonalizadoException for handling specific withdrawal errors.
- Created RetiroPersonalizadoNominaService to manage personalized withdrawal logic, including options retrieval and record reversal.
- Developed modal view for personalized withdrawal in the Blade template.
- Implemented comprehensive tests for the withdrawal service, covering various scenarios including manual concept withdrawal, quota and loan reversals, and validation of filter requirements.

+++ appsiel/app/Nomina/NomDocEncabezado.php
@@ -76,8 +76,8 @@ class NomDocEncabezado extends Model
 
     public function actualizar_totales()
     {
-        $this->total_devengos = NomDocRegistro::where('nom_doc_encabezado_id',$this->id)->sum('valor_devengo');
-        $this->total_deducciones = NomDocRegistro::where('nom_doc_encabezado_id',$this->id)->sum('valor_deduccion');
+        $this->total_devengos = (float) NomDocRegistro::where('nom_doc_encabezado_id',$this->id)->sum('valor_devengo');
+        $this->total_deducciones = (float) NomDocRegistro::where('nom_doc_encabezado_id',$this->id)->sum('valor_deduccion');
         $this->save();
     }
